fix: address external-review findings — reorg cref residue, ILogSink conformance, taxonomy clarification
Four XML doc crefs broken by the namespace moves are fully qualified
(doc generation is off, so no build warning fired). ILogSink splits out
of Log.cs into its conformant FanaBridge.Core.Logging home, shrinking
the root-namespace exception to exactly the physics-required Log type.
docs/architecture.md now states explicitly that domain folders are
navigational taxonomy, not directional boundaries — assembly references
are the only compiler-enforced direction.
