+++ src/FanaBridge.Core/Logging/Log.cs
@@ -1,20 +1,9 @@
 // Exception: namespace FanaBridge (not FanaBridge.Core.Logging) so unqualified Log.* from both assemblies resolves via the shared root ancestor.
 #pragma warning disable IDE0130 // Log must live in root FanaBridge so unqualified Log.* resolves from both Core and plugin
+using FanaBridge.Core.Logging;
+
 namespace FanaBridge
 {
-    /// <summary>
-    /// Log output contract for the SimHub-free core. The plugin shell installs a
-    /// sink that forwards to SimHub's logger; standalone hosts (unit tests) leave
-    /// it unset and core logging is a silent no-op.
-    /// </summary>
-    public interface ILogSink
-    {
-        void Info(string message);
-        void Warn(string message);
-        void Error(string message);
-        void Debug(string message);
-    }
-
     /// <summary>
     /// Process-wide log seam for the core device stack. Kept static (not
     /// injected) because logging is the one cross-cutting concern threaded
