+++ src/FanaBridge.Core/Transport/IHidBusEnumerator.cs
@@ -34,7 +34,7 @@ namespace FanaBridge.Core.Transport
 
     /// <summary>
     /// Seam over HID bus enumeration (<c>DeviceList.Local</c> is a process-global
-    /// static that cannot be faked). <see cref="FanatecWheelbase"/> consumes this
+    /// static that cannot be faked). <see cref="FanaBridge.Core.Devices.FanatecWheelbase"/> consumes this
     /// for discovery so base-PID selection — which has historically bitten on
     /// interface-shape edge cases — is table-testable.
     /// </summary>
